refactor workerfarm filestructure and fix error handling of ipc
